requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pydantic==2.9.2
Pillow==10.4.0
numpy==1.26.4
pytesseract==0.3.13